Improved NDSL examples
- Use boilerplate from ndsl.boilerplate
- Fix typos & broken links in notebooks
- Added README with quickstart guide
