+++ examples/NDSL/README.md
@@ -1,0 +1,14 @@
+# NDSL examples
+
+This folder contains a couple of Jupyter notebooks with the following examples:
+
+- Getting started with GT4Py: [GT4Py basics](./01_gt4py_basics.ipynb)
+- Getting started with NDSL middleware: [NDSL basics](./02_NDSL_basics.ipynb)
+- Combining stencil and non-stencil code with DaCe: [Orchestration basics](./03_orchestration_basics.ipynb)
+
+## Quickstart
+
+1. Make sure you fulfill the [requirements of NDSL](../../README.md#quickstart), e.g. python version.
+2. Create a virtual environment `python -m venv .venv`, and activate it `source .venv/bin/activate`.
+3. Install NDSL into your environment `pip install ../../[demos]`.
+4. In VSCode, install the Jupyter extension, select your virtual environment as kernel and run the notebooks.
